packages/fastsmtp/pyproject.toml
[project]
name = "fastsmtp"
version = "0.1.0"
description = "TLS-capable async SMTP server that receives emails, validates DKIM/SPF, and forwards contents to configurable webhooks"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "pydantic-settings>=2.6",
    "typer>=0.14",
    "httpx>=0.28",
    "aiosmtpd>=1.4",
    "sqlalchemy[asyncio]>=2.0",
    "alembic>=1.14",
    "tenacity>=9.0",
    "dkimpy>=1.1",
    "pyspf>=2.0",
    "dnspython>=2.7",
    "asyncpg>=0.30",
]

[project.optional-dependencies]
mariadb = ["aiomysql>=0.2"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "aiosmtplib>=3.0",
    "ruff>=0.8",
    "coverage>=7.0",
]

[project.scripts]
fastsmtp = "fastsmtp.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/fastsmtp"]
